#115 refactor: 게시글 등록 페이지 Select 컴포넌트 등록
