+++ src/pages/communityPage/category/CommunityPicture.jsx
