--- src/pages/communityPage/writePage/CommunityWritePage.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function CommunityWritePage() {
-  return <div>writePage</div>;
-}
